Added assignment cards and style
The navigation style has been updated. Assignment cards have been added.
Each card has the assignments for a specific subject.
Now, it occurs to me that giving each assignment a card of its own is a more
sensible idea. That should be the next goal. The layout of the cards, and the
animations must also be finalized.
A link must be added that let the user create a new assignment.
Login page needs work.
Homepage design must be finalized.
WE HAVE A LOT OF WORK TO DO.