Changed Assignments format, added a signup page and added our names
I have made the Assignment cards to refer to one particular assignment
rather than to all th eassignments for a subject.
I have added our names at the bottom of the index page.
I have linked the signup page to the button on the index page.
We still need to work on a filter for the assignments page.