Added New Assignment page, finalized login and signup
I have added a button for creating a new Assignment on the
Assignments page, and a page to create a new assignment, style and all.
The Login and Sign Up pages have been finalized.
We may have to drop the "Storage" idea. It doesn't make sense.
We can add links while creating the assignment to relevant materials.